% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{build_between_within_distributions}
\alias{build_between_within_distributions}
\title{Sample to construct between- and within-host distributions of correlation
for a given pair of taxa. The "within" distribution characterizes correlation
of a pair of taxa within selected hosts; the "between" distribution
characterizes the correlation of a given taxon with its own series across
a samples of hosts.}
\usage{
build_between_within_distributions(
  pair,
  coord1,
  coord2,
  predictions,
  selected_hosts,
  verbose = TRUE
)
}
\arguments{
\item{pair}{index of taxon pair}

\item{coord1}{index of taxon 1}

\item{coord2}{index of taxon 2}

\item{predictions}{predictions data.frame subsetted to the taxa of interest}

\item{selected_hosts}{subset of full host short name list to use}

\item{verbose}{if TRUE, prints status comments}
}
\value{
data.frame of between- and within- correlation distribution samples
}
\description{
Sample to construct between- and within-host distributions of correlation
for a given pair of taxa. The "within" distribution characterizes correlation
of a pair of taxa within selected hosts; the "between" distribution
characterizes the correlation of a given taxon with its own series across
a samples of hosts.
}
