% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc_consensus_sign}
\alias{calc_consensus_sign}
\title{Calculate a consensus CLR correlation sign for a given association pair}
\usage{
calc_consensus_sign(x)
}
\arguments{
\item{x}{vector of correlations between a pair of CLR microbes across hosts}
}
\value{
positive or negative consensus sign
}
\description{
Calculate a consensus CLR correlation sign for a given association pair
}
