% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc_universality_score}
\alias{calc_universality_score}
\title{Calculate a score (0.0 to 1.0) for the "universality" of given association
pair}
\usage{
calc_universality_score(x)
}
\arguments{
\item{x}{vector of correlations between a pair of CLR microbes across hosts}
}
\value{
numeric "score"
}
\description{
Calculate a score (0.0 to 1.0) for the "universality" of given association
pair
}
