% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc_universality_score_taxon}
\alias{calc_universality_score_taxon}
\title{Calculate a score (0.0 to 1.0) for the "universality" of association pairs
for each taxon}
\usage{
calc_universality_score_taxon(rug_obj, collapse = "unsigned")
}
\arguments{
\item{rug_obj}{output of the summarize_Sigmas function}

\item{collapse}{if value is "signed", computes an average
universality score for a taxon split up for consensus positive and negative
associations; if "unsigned" (default), computes na overall average universality score
for a taxon}
}
\value{
list of numeric "scores"
}
\description{
Calculate a score (0.0 to 1.0) for the "universality" of association pairs
for each taxon
}
