% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_dir}
\alias{check_dir}
\title{Create output directory/ies that may not already exist}
\usage{
check_dir(path_to_dir)
}
\arguments{
\item{path_to_dir}{a character vector of subdirectories to create (if not
already in existence)}
}
\value{
the path as a joined string
}
\description{
Create output directory/ies that may not already exist
}
