% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{convert_alr_Sigma_clr}
\alias{convert_alr_Sigma_clr}
\title{Convert an ALR covariance matrix to a CLR correlation matrix}
\usage{
convert_alr_Sigma_clr(Sigma)
}
\arguments{
\item{Sigma}{covariance matrix}
}
\value{
named list of CLR covariance and CLR correlation matrices
}
\description{
Convert an ALR covariance matrix to a CLR correlation matrix
}
