% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gLV_simulations.R
\name{convert_series_clr}
\alias{convert_series_clr}
\title{CLR-convert a set of species abundance series, calculate covariance across species, and return
the upper triangular portion of this covariance matrix}
\usage{
convert_series_clr(TS)
}
\arguments{
\item{TS}{multi-species time series returned by generate_series}
}
\value{
atomic vector
}
\description{
CLR-convert a set of species abundance series, calculate covariance across species, and return
the upper triangular portion of this covariance matrix
}
