% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{downsample_counts}
\alias{downsample_counts}
\title{Simulate count data from the generative model, downsampling it repeatedly,
and estimate the CLR correlation across taxa. Score the fidelity of estimates
on this downsampled data by calculating the proportion of taxon-taxon
correlations matched for sign.}
\usage{
downsample_counts(sample_no = c(10, 20, 30, 40, 50, 75, 100, 200))
}
\arguments{
\item{sample_no}{vector of sample counts; simulated data will be iteratively
downsampled to these sample counts}
}
\description{
Simulate count data from the generative model, downsampling it repeatedly,
and estimate the CLR correlation across taxa. Score the fidelity of estimates
on this downsampled data by calculating the proportion of taxon-taxon
correlations matched for sign.
}
