% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{filter_data}
\alias{filter_data}
\title{This function loads the raw ABRP data (as a phyloseq object) and 1) filters
to a set of best-sampled hosts, 2) agglomerates taxa to the level specified,
and 3) merges taxa below a specified minimum abundance across hosts}
\usage{
filter_data(
  tax_level = "ASV",
  host_sample_min = 75,
  count_threshold = 1,
  sample_threshold = 0.2,
  seq_similarity_threshold = 0.99
)
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{host_sample_min}{minimum sample number for host inclusion in the
filtered data set}

\item{count_threshold}{minimum count for taxon inclusion in the filtered data
set}

\item{sample_threshold}{minimum proportion of samples within each host at
which a taxon must be observed at or above count_threshold}
}
\value{
phyloseq object
}
\description{
This function loads the raw ABRP data (as a phyloseq object) and 1) filters
to a set of best-sampled hosts, 2) agglomerates taxa to the level specified,
and 3) merges taxa below a specified minimum abundance across hosts
}
\details{
Together count_threshold and sample_threshold specify a minimum
representation for a taxon. Taxa below this threshold will be grouped
together into an <NA> category.
}
