% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{filter_joint_zeros}
\alias{filter_joint_zeros}
\title{Return a vector indicating whether the frequency of 0-0 observations for a
given taxon pair is less than some threshold}
\usage{
filter_joint_zeros(counts, threshold_and = 0.05, threshold_or = 0.5)
}
\arguments{
\item{counts}{taxon-by-sample count matrix}

\item{threshold_and}{ceiling on the proportion of allowable joint zero (0-0)
observations for a given pair of taxa}

\item{threshold_or}{ceiling on the proportion of allowable combined zero and
non-zero observations for a given pair of taxa}
}
\value{
logical vector
}
\description{
Return a vector indicating whether the frequency of 0-0 observations for a
given taxon pair is less than some threshold
}
