% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_DLM}
\alias{fit_DLM}
\title{Fit MLN dynamic linear model to a given host's series using fido}
\usage{
fit_DLM(sname, counts, metadata, point_est = TRUE, smoothed = FALSE)
}
\arguments{
\item{sname}{host short name indicating which baboon's series to fit}

\item{counts}{filtered 16S count table (taxa x samples)}

\item{metadata}{annotations data.frame}

\item{point_est}{flag indicating whether or not to return (single) MAP
estimate}

\item{smoothed}{flag indicating whether or not to apply the simulation
smoother (necessary for trajectory visualization only)}
}
\value{
fidofit object
}
\description{
Fit MLN dynamic linear model to a given host's series using fido
}
