% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gLV_simulations.R
\name{generate_env_link}
\alias{generate_env_link}
\title{Generate matrix that encodes per-species response to environmental perturbation
This code is adapted directly from Tredennick et al. (2017)}
\usage{
generate_env_link(S, beta)
}
\arguments{
\item{S}{number of species to simulate}

\item{beta}{parameter returned by generate_innate_params}
}
\value{
matrix
}
\description{
Generate matrix that encodes per-species response to environmental perturbation
This code is adapted directly from Tredennick et al. (2017)
}
