% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gLV_simulations.R, R/visuals.R
\name{generate_highcontrast_palette}
\alias{generate_highcontrast_palette}
\title{Generate a random high-contrast palette (for bar plotting)}
\usage{
generate_highcontrast_palette(S)

generate_highcontrast_palette(S)
}
\arguments{
\item{S}{number of features (colors to generate)}
}
\value{
vector of hex color strings

list of hex colors
}
\description{
Generate a random high-contrast palette (for bar plotting)

Generate a color palette suitable for stacked bar plots given a feature number (S)
}
