% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{generate_highcontrast_palette2}
\alias{generate_highcontrast_palette2}
\title{Generate a color palette suitable for stacked bar plots given a feature number (S)
This version uses the Polychrome package in R and (I think) gives a more
visually parsable result}
\usage{
generate_highcontrast_palette2(
  S,
  seed_colors_hex = c("#ff0000", "#00ff00", "#0000ff")
)
}
\arguments{
\item{S}{number of features (colors to generate)}

\item{seed_colors_hex}{optional list of hex colors to seed function with}
}
\value{
list of hex colors
}
\description{
Generate a color palette suitable for stacked bar plots given a feature number (S)
This version uses the Polychrome package in R and (I think) gives a more
visually parsable result
}
