% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gLV_simulations.R
\name{generate_innate_params}
\alias{generate_innate_params}
\title{Generate baseline abundances and cooperative/competitive parameters for use a generalized
stochastic Lotka-Volterra
This code is adapted directly from Tredennick et al. (2017)}
\usage{
generate_innate_params(S, start_time, end_time, PhiE, almean)
}
\arguments{
\item{S}{number of species to simulate}

\item{start_time}{time point at which to begin "recording" the series (previous time points will be
discarded as burn-in)}

\item{end_time}{full length of series to simulate (including burn-in)}

\item{PhiE}{value from -1 to 1 defining average "synchrony" of species in the system;
-1 indicates strong asynchrony, 1 indicates strong synchrony, and 0, neutrality}

\item{almean}{value from -1 to 1 defining average competition or cooperation of species in
the system; I believe -1.0 indicates strong cooperation, 1.0 indicates strong competition}
}
\value{
named list of parameters
}
\description{
Generate baseline abundances and cooperative/competitive parameters for use a generalized
stochastic Lotka-Volterra
This code is adapted directly from Tredennick et al. (2017)
}
