% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{generate_palette}
\alias{generate_palette}
\title{Generate a color palette suitable for stacked bar plots given a feature number (S)}
\usage{
generate_palette(hex_min, hex_mid, hex_max, S)
}
\arguments{
\item{hex_min}{hex code for min color in ramp}

\item{hex_mid}{hex code for median color in ramp}

\item{hex_max}{hex code for max color in ramp}

\item{S}{number of colors to generate}
}
\value{
list of hex colors
}
\description{
Generate a color palette suitable for stacked bar plots given a feature number (S)
}
