% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{get_Gamma}
\alias{get_Gamma}
\title{Define a kernel (function) over samples}
\usage{
get_Gamma(
  kernel_scale,
  proportions,
  min_correlation = 0.1,
  days_to_baseline = 90
)
}
\arguments{
\item{kernel_scale}{total variance for the composite kernel}

\item{proportions}{proportion variance to attribute to each of 2 kernels (see
details)}

\item{rho}{bandwidth for SE kernel}
}
\value{
list containing kernel function and bandwidth parameter
}
\description{
Define a kernel (function) over samples
}
\details{
Composite kernel is built from (1) squared exponential kernel (base
autocorrelation component) and (2) seasonal kernel (periodic)
}
