% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{get_host_social_groups}
\alias{get_host_social_groups}
\title{Assign host social group as social group in which a given host has the
majority of its samples}
\usage{
get_host_social_groups(host_list = NULL)
}
\arguments{
\item{host_list}{optional list of host short names}
}
\value{
list of host short name to social group mappings
}
\description{
Assign host social group as social group in which a given host has the
majority of its samples
}
