% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_mean_clr_abundance}
\alias{get_mean_clr_abundance}
\title{Get the mean CLR abundance for all taxa}
\usage{
get_mean_clr_abundance(tax_level = "ASV")

get_mean_clr_abundance(tax_level = "ASV")
}
\arguments{
\item{tax_level}{taxonomic level at which to calculate average CLR abundance}
}
\value{
vector of mean CLR abundances

vector of mean CLR abundances
}
\description{
Get the mean CLR abundance for all taxa

Get the mean CLR abundance for all taxa
}
