% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{get_predictions_host_list}
\alias{get_predictions_host_list}
\title{Utility function for pulling collection date-aligned trajectories for a pair
of taxa from a set of hosts.}
\usage{
get_predictions_host_list(ref_hosts, metadata)
}
\description{
Utility function for pulling collection date-aligned trajectories for a pair
of taxa from a set of hosts.
}
