% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_tax_label}
\alias{get_tax_label}
\title{Return a human-readable taxon label}
\usage{
get_tax_label(tax, coord, coord_system_label)
}
\arguments{
\item{tax}{taxonomy data.frame}

\item{coord}{index of logratio to label}

\item{coord_system_label}{one of c("CLR", "ALR")}
}
\value{
label
}
\description{
Return a human-readable taxon label
}
