% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{get_trajectory_df}
\alias{get_trajectory_df}
\title{Utility function for pulling collection date-aligned trajectories from a
given host's prediction set.}
\usage{
get_trajectory_df(dates, coord, common_baseline, predictions, center = NULL)
}
\description{
Utility function for pulling collection date-aligned trajectories from a
given host's prediction set.
}
