% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{load_data}
\alias{load_data}
\title{This function wrangles the filtered ABRP data and metadata}
\usage{
load_data(
  tax_level = "ASV",
  host_sample_min = 75,
  count_threshold = 1,
  sample_threshold = 0.2,
  alr_median = FALSE
)
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{host_sample_min}{minimum sample number for host inclusion in the
filtered data set}

\item{count_threshold}{minimum count for taxon inclusion in the filtered data
set}

\item{sample_threshold}{minimum proportion of samples within each host at
which a taxon must be observed at or above count_threshold}

\item{alr_median}{if TRUE, uses the taxon with the median coefficient of 
variation in relative abundance as the ALR reference taxon, otherwise uses
the taxon in the last index in the count table}
}
\value{
a named list of count table, taxonomy, and metadata components
}
\description{
This function wrangles the filtered ABRP data and metadata
}
\details{
Together count_threshold and sample_threshold specify a minimum
representation for a taxon. Taxa below this threshold will be grouped
together into an <NA> category.
}
