% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{load_filtered_data}
\alias{load_filtered_data}
\title{This function loads the filtered ABRP data (as a phyloseq object)}
\usage{
load_filtered_data(
  tax_level = "ASV",
  host_sample_min = 75,
  count_threshold = 5,
  sample_threshold = 0.2,
  process_if_absent = FALSE
)
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{host_sample_min}{minimum sample number for host inclusion in the
filtered data set}

\item{count_threshold}{minimum count for taxon inclusion in the filtered data
set}

\item{sample_threshold}{minimum proportion of samples within each host at
which a taxon must be observed at or above count_threshold}

\item{process_if_absent}{flag indicating whether to process the raw data if
filtered data file is not found}
}
\value{
A named list of count table, taxonomy, and metadata components
}
\description{
This function loads the filtered ABRP data (as a phyloseq object)
}
\details{
Together count_threshold and sample_threshold specify a minimum
representation for a taxon. Taxa below this threshold will be grouped
together into an <NA> category.
}
