% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{load_scrambled_data}
\alias{load_scrambled_data}
\title{Load scrambled data}
\usage{
load_scrambled_data(
  tax_level = "ASV",
  host_sample_min = 75,
  count_threshold = 1,
  sample_threshold = 0.2
)
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{host_sample_min}{minimum sample number for host inclusion in the
filtered data set}

\item{count_threshold}{minimum count for taxon inclusion in the filtered data
set}

\item{sample_threshold}{minimum proportion of samples within each host at
which a taxon must be observed at or above count_threshold}
}
\value{
a named list of count table, taxonomy, and metadata components
}
\description{
Load scrambled data
}
