% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{order_rug_col_phylogenetic}
\alias{order_rug_col_phylogenetic}
\title{Reorder columns of the "rug" based on phylogenetic distance}
\usage{
order_rug_col_phylogenetic(rug_obj, taxonomy)
}
\arguments{
\item{rug_obj}{output of the summarize_Sigmas function}

\item{taxonomy}{taxonomy appropriate for data at the same level as the rug
object}
}
\value{
row ordering
}
\description{
Reorder columns of the "rug" based on phylogenetic distance
}
