% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{order_rug_cols_mean_abundance}
\alias{order_rug_cols_mean_abundance}
\title{Reorder columns of the "rug" from taxa with the least to most difference in
mean CLR abundance}
\usage{
order_rug_cols_mean_abundance(rug_obj, counts)
}
\arguments{
\item{rug_obj}{output of the summarize_Sigmas function}

\item{counts}{count table at the same taxonomic level as results in rug_obj}
}
\value{
named list of column ordering and absolute CLR differences
}
\description{
Reorder columns of the "rug" from taxa with the least to most difference in
mean CLR abundance
}
