% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{order_rug_row_baseline}
\alias{order_rug_row_baseline}
\title{Reorder rows of the "rug" based on Aitchison distance between host average
compositions}
\usage{
order_rug_row_baseline(rug_obj, counts, metadata)
}
\arguments{
\item{rug_obj}{output of the summarize_Sigmas function}

\item{counts}{count table at the same taxonomic level as results in rug_obj}

\item{metadata}{metadata at the same taxonomic level as results in rug_obj}
}
\value{
row ordering
}
\description{
Reorder rows of the "rug" based on Aitchison distance between host average
compositions
}
