% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{order_rug_row_diversity}
\alias{order_rug_row_diversity}
\title{Reorder rows of the "rug" based on average taxonomic average host alpha
diversity (low to high Shannon index)}
\usage{
order_rug_row_diversity(rug_obj)
}
\arguments{
\item{rug_obj}{output of the summarize_Sigmas function}
}
\value{
row ordering
}
\description{
Reorder rows of the "rug" based on average taxonomic average host alpha
diversity (low to high Shannon index)
}
