% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{PER}
\alias{PER}
\title{Periodic kernel}
\usage{
PER(X, sigma = 1, rho = 1, period = 24, jitter = 0)
}
\arguments{
\item{X}{covariate (dimension Q x N; i.e., covariates x samples)}

\item{sigma}{scalar parameter}

\item{rho}{scalar bandwidth parameter}

\item{period}{period length (in days)}

\item{jitter}{small scalar to add to off-diagonal of gram matrix
(for numerical underflow issues)}
}
\value{
Gram Matrix (N x N) (e.g., the Kernel evaluated at
each pair of points)
}
\description{
Periodic kernel
}
