% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{percent_to_k}
\alias{percent_to_k}
\title{Converts a percent to a relative count}
\usage{
percent_to_k(percent, n_features)
}
\arguments{
\item{percent}{percent}

\item{n_features}{total number of features}
}
\value{
relative count
}
\description{
Converts a percent to a relative count
}
