% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plot_aligned_trajectories}
\alias{plot_aligned_trajectories}
\title{Render date-aligned trajectories for a pair of taxa and a given set of hosts.
This allows us to eyeball whether taxa strongly correlated within hosts show
similar dynamics across hosts.}
\usage{
plot_aligned_trajectories(
  host_list,
  tax_idx1,
  tax_idx2,
  metadata,
  save_file = FALSE
)
}
\arguments{
\item{host_list}{vector of host short names}

\item{tax_idx1}{index of first taxon}

\item{tax_idx2}{index of second taxon}

\item{metadata}{metadata data.frame (with sname and collection_date column)}

\item{save_file}{flag indicating whether or not to save rendered plot to file}
}
\description{
Render date-aligned trajectories for a pair of taxa and a given set of hosts.
This allows us to eyeball whether taxa strongly correlated within hosts show
similar dynamics across hosts.
}
