% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{plot_clr_vs_diet}
\alias{plot_clr_vs_diet}
\title{Render line plot of a given CLR taxon against diet PC1 for a given host.
This plot is meant to demonstrate the largely uncorrelated nature of the diet
data with the observed logratio abundances.}
\usage{
plot_clr_vs_diet(sname, tax_idx, counts, metadata)
}
\arguments{
\item{sname}{host short name indicating which baboon's series to fit}

\item{tax_idx}{symmetric matrix object}

\item{counts}{filtered 16S count table (taxa x samples)}

\item{metadata}{annotations data.frame}
}
\description{
Render line plot of a given CLR taxon against diet PC1 for a given host.
This plot is meant to demonstrate the largely uncorrelated nature of the diet
data with the observed logratio abundances.
}
