% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plot_correlation_histogram}
\alias{plot_correlation_histogram}
\title{Renders the host x pairwise associations as a histogram}
\usage{
plot_correlation_histogram(rug)
}
\arguments{
\item{rug}{heatmap matrix output from summarize_Sigmas()}

\item{save_name}{name with which to save heatmap file}
}
\value{
named list with column and row ordering
}
\description{
Renders the host x pairwise associations as a histogram
}
