% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{plot_downsampling}
\alias{plot_downsampling}
\title{Boxplot data simulated from downsample_counts()}
\usage{
plot_downsampling(data)
}
\arguments{
\item{data}{data.frame containing results from downsample_counts()}
}
\description{
Boxplot data simulated from downsample_counts()
}
