% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plot relative representation of ASV pairs (using family labels)}
\usage{
plot_enrichment(
  frequencies_subset,
  frequencies,
  plot_height,
  plot_width,
  legend_topmargin,
  legend_leftmargin,
  save_name
)
}
\arguments{
\item{frequencies_subset}{table of ASV pairs (labeled as "Family - Family")
and their frequency in some subset of case of interest}

\item{frequencies}{table of ASV pairs (labeled as "Family - Family")
and their pverall frequency}

\item{plot_height}{plot height in inches}

\item{plot_width}{plot width in inches}

\item{legend_topmargin}{top margin of legend in points}

\item{legend_leftmargin}{left margin of legend in inches}
}
\description{
Plot relative representation of ASV pairs (using family labels)
}
