% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gLV_simulations.R
\name{plot_gLV_bars}
\alias{plot_gLV_bars}
\title{Plot a random 50-day segment from a gLV simulation}
\usage{
plot_gLV_bars(series, palette = NULL)
}
\arguments{
\item{series}{matrix of observations from a single host (rows are species, columns are time points)}

\item{palette}{optional color palette (as a vector of hex color strings)}
}
\description{
Plot a random 50-day segment from a gLV simulation
}
