% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gLV_simulations.R
\name{plot_gLV_heatmap}
\alias{plot_gLV_heatmap}
\title{Plot a gLV simulation as a heatmap}
\usage{
plot_gLV_heatmap(heatmap)
}
\arguments{
\item{heatmap}{matrix of correlations (rows are hosts, columns are taxon-taxon pairs)}
}
\description{
Plot a gLV simulation as a heatmap
}
