% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plot_rug}
\alias{plot_rug}
\title{Renders the "rug" (host x pairwise association matrix) as a heatmap.}
\usage{
plot_rug(
  rug,
  canonical_col_order = NULL,
  canonical_row_order = NULL,
  row_labels = NULL,
  cluster_obj = NULL,
  save_name = NULL
)
}
\arguments{
\item{rug}{heatmap matrix output from summarize_Sigmas()}

\item{canonical_col_order}{if not NULL, order of pairs (columns) to use}

\item{canonical_row_order}{if not NULL, order of rows (hosts) to use}

\item{row_labels}{optional parameter with row labels (ordered host short
names)}

\item{cluster_obj}{an optional pre-computed hierarchical clustering; if
included, a dendrogram will be rendered from this}

\item{save_name}{name with which to save heatmap file}
}
\value{
named list with column and row ordering
}
\description{
Renders the "rug" (host x pairwise association matrix) as a heatmap.
}
