% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plot_trajectories_top_pairs}
\alias{plot_trajectories_top_pairs}
\title{Render date-aligned trajectories for the most universal positive and negative
pairs of taxa in the top-sampled hosts within each social group}
\usage{
plot_trajectories_top_pairs(output_dir, metadata, taxonomy)
}
\arguments{
\item{output_dir}{specifies the subdirectory of the model fits directory in
which to save the predictions}

\item{metadata}{metadata data.frame (with sname and collection_date column)}

\item{taxonomy}{appropriate taxonomy object for this level}
}
\description{
Render date-aligned trajectories for the most universal positive and negative
pairs of taxa in the top-sampled hosts within each social group
}
