% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Visualize inferred logratio trajectories. GP objects will be visualized in
the CLR coordinate system; DLM objects will be visualized in the ALR (for
now).}
\usage{
plot_trajectory(
  fit,
  coord,
  coord_label = NULL,
  show_observations = FALSE,
  save_file = FALSE
)
}
\arguments{
\item{fit}{bassetfit or labraduckfit object (from fido)}

\item{coord}{index of logratio to plot}

\item{coord_label}{optional coord label (e.g. "family Lachnospiraceae")}

\item{show_observations}{flag indicating whether or not to logratio transform
and render the observed data; this can be useful for judging sensitivity of
fit}

\item{save_file}{flag indicating whether or not to save rendered plot to file}
}
\description{
Visualize inferred logratio trajectories. GP objects will be visualized in
the CLR coordinate system; DLM objects will be visualized in the ALR (for
now).
}
