% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{predict_GP_mean}
\alias{predict_GP_mean}
\title{Predict mean trajectories (Lambda) for all CLR taxa in a given host}
\usage{
predict_GP_mean(
  fit = NULL,
  host = NULL,
  output_dir = NULL,
  interpolation = "linear"
)
}
\arguments{
\item{fit}{an optional fitted model object}

\item{host}{an optional host short name}

\item{output_dir}{an optional model fit directory}

\item{interpolation}{allowed values are "mean", "linear", or "none"}
}
\value{
named list with inferred trajectories and time span
}
\description{
Predict mean trajectories (Lambda) for all CLR taxa in a given host
}
\details{
Either `fit` or `host` and `output_dir` must be provided. If all are
provided, `fit` overrides all. Note also: the runtime on this function is
around 10s per host at the ASV-level.
}
