% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{predict_span}
\alias{predict_span}
\title{Draw predictions from a bassetfit or labraduckfit object over a given day
range}
\usage{
predict_span(fit, span = NULL)
}
\arguments{
\item{fit}{bassetfit or labraduckfit object (from fido)}

\item{span}{optional day range over which to predict}
}
\description{
Draw predictions from a bassetfit or labraduckfit object over a given day
range
}
