% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{predict_trajectory}
\alias{predict_trajectory}
\title{Draw predictions from a fido::basset model in the CLR coordinate system.
This function interpolates 5% of missing observations, which gives good quick
visuals in the ABRP data.}
\usage{
predict_trajectory(fit)
}
\arguments{
\item{fit}{bassetfit object}
}
\description{
Draw predictions from a fido::basset model in the CLR coordinate system.
This function interpolates 5% of missing observations, which gives good quick
visuals in the ABRP data.
}
