% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{prune_data}
\alias{prune_data}
\title{Description TBD}
\usage{
prune_data(seq_similarity_threshold = 0.99)
}
\arguments{
\item{seq_similarity_threshold}{percent upper limit on sequence similarity;
for ASV pairs strictly more similar than this, the less abundant partner in
the pair will be removed from the data as it's likely these partners ID the
same taxon}

\item{host_sample_min}{minimum sample number for host inclusion in the
filtered data set}

\item{count_threshold}{minimum count for taxon inclusion in the filtered data
set}

\item{sample_threshold}{minimum proportion of samples within each host at
which a taxon must be observed at or above count_threshold}
}
\value{
phyloseq object
}
\description{
Description TBD
}
\details{
Typically, ASV pairs with >99% similarity differ by 1-2 bases only
}
