% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pull_Sigmas}
\alias{pull_Sigmas}
\title{Gather all host ASV-ASV (MAP) correlation matrices into a 3D array}
\usage{
pull_Sigmas(output_dir)
}
\arguments{
\item{output_dir}{specifies the subdirectory of the model fits directory in
which to look for fitted model output}
}
\value{
D x D x H array of host ASV-ASV correlation matrices
}
\description{
Gather all host ASV-ASV (MAP) correlation matrices into a 3D array
}
