% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{render_universal_pairs}
\alias{render_universal_pairs}
\title{Utility function to plot aligned series for the pair of taxa with the largest
universality score from among a subset of association indices}
\usage{
render_universal_pairs(
  output_dir,
  select_idx,
  scores,
  tax_idx1,
  tax_idx2,
  taxonomy
)
}
\description{
Utility function to plot aligned series for the pair of taxa with the largest
universality score from among a subset of association indices
}
