% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rug_phylogenetic_distances}
\alias{rug_phylogenetic_distances}
\title{Calculate phylogenetic distances on pairs of taxa}
\usage{
rug_phylogenetic_distances(rug_obj, taxonomy, as_matrix = FALSE)
}
\arguments{
\item{rug_obj}{output of the summarize_Sigmas function}

\item{taxonomy}{taxonomy appropriate for data at the same level as the rug
object}

\item{as_matrix}{if TRUE, returns the distance matrix, not the vectorized
result}
}
\value{
row ordering
}
\description{
Calculate phylogenetic distances on pairs of taxa
}
