% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity_sweep}
\alias{sensitivity_sweep}
\title{Use the "rug" plot to visualize the output of multiple model runs. If no
column/row orderings are specified, all plots will be clustered according
to the column/row orderings derived from the first set of models specified.}
\usage{
sensitivity_sweep(
  output_dir_list,
  canonical_col_order = NULL,
  canonical_row_order = NULL
)
}
\arguments{
\item{output_dir_list}{specifies the subdirectories of the model fits
directory in which to look for fitted model output}

\item{canonical_col_order}{if not NULL, order of pairs (columns) to use}

\item{canonical_row_order}{if not NULL, order of rows (hosts) to use}
}
\description{
Note: This ASSUMES method of interest is the Gaussian process.
}
