% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sequence_distance}
\alias{sequence_distance}
\title{Compute sequence distances across ASVs}
\usage{
sequence_distance(distance_type = "N")
}
\arguments{
\item{distance_type}{optional; default is "N" (number of mismatches)}
}
\value{
matrix of ASV-ASV distances in terms of their 16S sequences
}
\description{
Compute sequence distances across ASVs
}
