% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_series}
\alias{simulate_series}
\title{Simulate series for one host}
\usage{
simulate_series(D = 100, N = 1000, Sigma = NULL, Theta = NULL)
}
\arguments{
\item{D}{number of features}

\item{N}{number of samples (time points)}

\item{Sigma}{optional correlation matrix to simulate from}

\item{Theta}{optional baseline for logratios}
}
\description{
Simulate series for one host
}
