% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_GP}
\alias{fit_GP}
\title{Fit MLN Gaussian process model to a given host's series using fido}
\usage{
fit_GP(sname, counts, metadata, point_est = TRUE)
}
\arguments{
\item{sname}{host short name indicating which baboon's series to fit}

\item{counts}{filtered 16S count table (taxa x samples)}

\item{metadata}{annotations data.frame}

\item{point_est}{flag indicating whether or not to return (single) MAP
estimate}
}
\value{
fidofit object
}
\description{
Fit MLN Gaussian process model to a given host's series using fido
}
