% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_GP}
\alias{fit_GP}
\title{Fit MLN Gaussian process model to a given host's series using fido}
\usage{
fit_GP(
  sname,
  counts,
  metadata,
  output_dir,
  MAP = TRUE,
  days_to_min_autocorrelation = 90,
  diet_weight = 0,
  var_scale_taxa = 1,
  var_scale_samples = 1,
  use_adam = FALSE
)
}
\arguments{
\item{sname}{host short name indicating which baboon's series to fit}

\item{counts}{filtered 16S count table (taxa x samples)}

\item{metadata}{annotations data.frame}

\item{output_dir}{specifies the subdirectory of the model fits directory in
which to save the fitted model output}

\item{MAP}{flag indicating whether or not to return (single) MAP estimate}

\item{days_to_min_autocorrelation}{used in the calculation of bandwidth for
the kernel over samples that models autocorrelation; indicates the number of
days at which the squared exponential correlation decays to a minimum (here,
0.1)}

\item{diet_weight}{relative contribution of first three diet PCs to
covariance across samples}

\item{var_scale_taxa}{scale of the prior associated with the covariance over
taxa}

\item{var_scale_samples}{scale of the hyperparameter associated with the
covariance over samples}

\item{use_adam}{optimize with Adam (occasionally this converges more reliably
that default L-BFGS)}
}
\value{
fidofit object
}
\description{
Fit MLN Gaussian process model to a given host's series using fido
}
