% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_GP}
\alias{fit_GP}
\title{Fit MLN Gaussian process model to a given host's series using fido}
\usage{
fit_GP(
  sname,
  counts,
  metadata,
  point_est = TRUE,
  output_dir = NULL,
  diet_weight = 0,
  days_to_min_autocorrelation = 90
)
}
\arguments{
\item{sname}{host short name indicating which baboon's series to fit}

\item{counts}{filtered 16S count table (taxa x samples)}

\item{metadata}{annotations data.frame}

\item{point_est}{flag indicating whether or not to return (single) MAP
estimate}

\item{output_dir}{if supplied, specifies the subdirectory of the model fits
directory in which to save the fitted model output}

\item{diet_weight}{relative contribution of first three diet PCs to
covariance across samples}
}
\value{
fidofit object
}
\description{
Fit MLN Gaussian process model to a given host's series using fido
}
