% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{generate_highcontrast_palette}
\alias{generate_highcontrast_palette}
\title{Generate a color palette suitable for stacked bar plots given a feature number (S)}
\usage{
generate_highcontrast_palette(S)
}
\arguments{
\item{S}{number of features (colors to generate)}
}
\value{
list of hex colors
}
\description{
Generate a color palette suitable for stacked bar plots given a feature number (S)
}
