% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{get_paired_trajectories}
\alias{get_paired_trajectories}
\title{Utility function for pulling collection date-aligned trajectories for a pair
of taxa from a given host's prediction set.}
\usage{
get_paired_trajectories(
  dates,
  coord1,
  coord2,
  host,
  common_baseline,
  predictions,
  center = NULL
)
}
\description{
Utility function for pulling collection date-aligned trajectories for a pair
of taxa from a given host's prediction set.
}
