% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{get_trajectory_df}
\alias{get_trajectory_df}
\title{Pull collection date-aligned trajectories from a given host's prediction set}
\usage{
get_trajectory_df(
  dates,
  coord,
  common_baseline,
  predictions,
  center = NULL,
  mean_only = FALSE
)
}
\arguments{
\item{dates}{named list of dates in YYYY-MM-DD format indexed by host short
name}

\item{coord}{index of taxon}

\item{common_baseline}{earliest date to use for all hosts (YYYY-MM-DD format)}

\item{predictions}{named list of predicted (interpolated) Eta matrices
indexed by host short name}

\item{center}{center to use for taxa (e.g. 0)}

\item{mean_only}{return mean rather than quantiles}
}
\description{
Pull collection date-aligned trajectories from a given host's prediction set
}
