% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plot_aligned_trajectories}
\alias{plot_aligned_trajectories}
\title{Render date-aligned trajectories for a pair of taxa and a given set of hosts.
This allows us to eyeball whether taxa strongly correlated within hosts show
similar dynamics across hosts.}
\usage{
plot_aligned_trajectories(
  output_dir,
  tax_idx1,
  tax_idx2,
  tax_label1,
  tax_label2,
  metadata,
  bind_taxa = FALSE,
  save_file = FALSE
)
}
\arguments{
\item{output_dir}{specifies the subdirectory of the model fits directory in
which to save the predictions}

\item{tax_idx1}{index of first taxon}

\item{tax_idx2}{index of second taxon}

\item{tax_label1}{readable label for first taxon}

\item{tax_label2}{readable label for second taxon}

\item{metadata}{metadata data.frame (with sname and collection_date column)}

\item{bind_taxa}{plot all trajectories for taxon 1 together; ditto taxon 2}

\item{save_file}{flag indicating whether or not to save rendered plot to file}
}
\description{
Render date-aligned trajectories for a pair of taxa and a given set of hosts.
This allows us to eyeball whether taxa strongly correlated within hosts show
similar dynamics across hosts.
}
