% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plot relative representation of ASV pairs (using family labels)}
\usage{
plot_enrichment(
  frequencies,
  frequencies_subset,
  type_label,
  location_label,
  enrichment = NULL,
  count_threshold = 10,
  cap_size = 5,
  pt_sz = 3,
  title_text_sz = 12,
  text_sz = 10,
  stroke_sz = 1.25
)
}
\arguments{
\item{frequencies}{table of ASV pairs (labeled as "Family - Family")
and their pverall frequency}

\item{frequencies_subset}{table of ASVs or ASV pairs (labeled as
"Family - Family") and their frequency in some subset of case of interest}

\item{type_label}{either "family" or "family-pair"}

\item{location_label}{a label for the subset, e.g. "high universality, low
synchrony taxon pairs"}

\item{enrichment}{optional data.frame onto which enrichment results will be
appended}

\item{count_threshold}{minimum abundance at which to evaluate taxon families
or family pairs for enrichment}

\item{cap_size}{log2 minimum or maximum value to plot; plotted intervals will
be truncated here}

\item{pt_sz}{optional geom_point size}

\item{title_text_sz}{optional axis title text size}

\item{text_sz}{optional axis text size}

\item{stroke_sz}{optional geom_linerange stroke size}
}
\value{
list containing the plot and enrichment data.frame
}
\description{
Plot relative representation of ASV pairs (using family labels)
}
