% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Visualize inferred logratio trajectories}
\usage{
plot_trajectory(
  fit,
  predictions,
  taxon,
  taxon_label = NULL,
  show_observations = FALSE,
  save_file = FALSE
)
}
\arguments{
\item{fit}{bassetfit or labraduckfit object (from fido)}

\item{predictions}{named list of predictions from a bassetfit or
labraduckfit object}

\item{taxon}{index of taxon to plot}

\item{taxon_label}{optional taxon label (e.g. "family Lachnospiraceae")}

\item{show_observations}{flag indicating whether or not to logratio transform
and render the observed data; this can be useful for judging sensitivity of
fit}

\item{save_file}{flag indicating whether or not to save rendered plot to file}
}
\description{
Visualize inferred logratio trajectories
}
