% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Visualize inferred logratio trajectories (CLR by default)}
\usage{
plot_trajectory(
  sname,
  output_dir,
  coord,
  coord_label = NULL,
  show_observations = FALSE,
  save_file = FALSE
)
}
\arguments{
\item{sname}{host short name indicating which baboon's series to fit}

\item{output_dir}{specifies the subdirectory of the model fits directory in
which to save the predictions}

\item{coord}{index of logratio to plot}

\item{coord_label}{optional coord label (e.g. "family Lachnospiraceae")}

\item{show_observations}{flag indicating whether or not to logratio transform
and render the observed data; this can be useful for judging sensitivity of
fit}

\item{save_file}{flag indicating whether or not to save rendered plot to file}
}
\description{
Visualize inferred logratio trajectories (CLR by default)
}
