% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{predict_GP_mean}
\alias{predict_GP_mean}
\title{Predict mean trajectories (Lambda) for all CLR taxa in a given host}
\usage{
predict_GP_mean(output_dir, host)
}
\arguments{
\item{output_dir}{model fit directory}

\item{host}{host short name}
}
\value{
named list with inferred trajectories and time span
}
\description{
Predict mean trajectories (Lambda) for all CLR taxa in a given host
}
\details{
The runtime on this function is around 10s per host at the ASV-level
}
