% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{predict_trajectory}
\alias{predict_trajectory}
\title{Draw predictions from a bassetfit or labraduckfit object over a given day
range. These will be in the CLR coordinate system!}
\usage{
predict_trajectory(fit, resolution = 100)
}
\arguments{
\item{fit}{bassetfit or labraduckfit object (from fido)}

\item{resolution}{percent of days to predict (e.g. 50 predicts half the
missing days/samples)}
}
\description{
Draw predictions from a bassetfit or labraduckfit object over a given day
range. These will be in the CLR coordinate system!
}
