% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{summarize_Sigmas}
\alias{summarize_Sigmas}
\title{Summarize the pairwise correlations (optionally via proportionality) from all
fitted models of a given type. Uses MAP estimates if available, otherwise,
summarizes full posterior model fits to give a "rug" matrix where rows are
hosts and columns are unique pairwise interactions between taxa.}
\usage{
summarize_Sigmas(output_dir, use_proportionality = FALSE)
}
\arguments{
\item{output_dir}{specifies the subdirectory of the model fits directory in
which to look for fitted model output}

\item{use_proportionality}{flag indicating whether to compute proportionality
between features; if FALSE, CLR correlations are computed}
}
\value{
named list of "rug" and associated labels
}
\description{
Summarize the pairwise correlations (optionally via proportionality) from all
fitted models of a given type. Uses MAP estimates if available, otherwise,
summarizes full posterior model fits to give a "rug" matrix where rows are
hosts and columns are unique pairwise interactions between taxa.
}
